Type Alias pallet_sponsorship::pallet::Module

pub type Module<T> = Pallet<T>;

👎Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T: Config> Pallet<T>

pub fn create_pot( origin: OriginFor<T>, pot: T::PotId, sponsorship_type: T::SponsorshipType, fee_quota: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance, reserve_quota: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance ) -> DispatchResult

Create a new sponsorship pot and set its limits. The pot id shouldn’t be in use.

Emits PotCreated(pot) event when successful.

pub fn remove_pot(origin: OriginFor<T>, pot: T::PotId) -> DispatchResult

Allows the sponsor to remove the pot they have created themselves. The pot must not have any users. Users must have been removed prior to this call.

Emits PotRemoved(pot) when successful

pub fn register_users( origin: OriginFor<T>, pot: T::PotId, users: Vec<T::AccountId>, common_fee_quota: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance, common_reserve_quota: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance ) -> DispatchResult

Register users for a pot and set the same limit for the list of them. Only pot sponsor can do this.

Emits UsersRegistered(pot, Vec<T::AccountId>) with a list of registered when successful.

pub fn remove_users( origin: OriginFor<T>, pot: T::PotId, users: Vec<T::AccountId> ) -> DispatchResult

Remove users from a pot. Only pot sponsor can do this. None of the specified users must have any reserved balance in their proxy accounts. User must be registered to be removable. Users receive the free balance in their proxy account back into their own accounts when they are removed.

Emits UsersRemoved(pot, Vec<T::AccountId>) with a list of those removed when successful.

pub fn sponsor_for( origin: OriginFor<T>, pot: T::PotId, call: Box<<T as Config>::RuntimeCall> ) -> DispatchResultWithPostInfo

Sponsor me for the given call from the specified pot. The caller must be registered for the pot. The call must be consistent with the pot’s sponsorship type.

Returns Error if the pot doesn’t exist or the user is not registered for the pot or if their call is not matching the sponsorship type in which case the error would be frame_system::Error::CallFiltered. Also returns error if the call itself should fail for any reason related to either the call or the available fund for the user. In this case the actual error will be depending on the call itself.
Regardless of the sponsorship type, users are not allowed to dispatch calls that would leak fund from their proxy account. If they try to do so they will get Error::BalanceLeak. For example they cannot transfer fund to another account even if the sponsorship type allows Balances calls.

Emits Sponsored {paid, repaid} when successful. The paid is the amount initially transferred to the proxy account of the user by the sponsor. The repaid is the amount repaid to the sponsor after the call has been successfully executed. Please note repaid can be bigger than paid if for any reason the user is able to partially or fully pay back their previous debt to the sponsor too. Also the paid might be less than what the limit for the user allows if the user can support themselves partially or fully based on their free balance in their proxy account . Finally, the paid is limited by the remaining reserve quota for the pot too.

Note: The addition of T::DbWeight::get().reads_writes(2, 2) to the weight is to account for the reads and writes of the pot_details and user_details storage items which are needed during pre and post dispatching this call.

pub fn update_pot_limits( origin: OriginFor<T>, pot: T::PotId, new_fee_quota: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance, new_reserve_quota: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance ) -> DispatchResult

Update the pot details. Only the sponsor can do this. If the sponsor is lowering their support, it can work only if the corresponding fee or reserve balance has enough available margin. In other words, the sponsor cannot lower the limit for the fee below what users have already taken from the pot. Similarly, the sponsor cannot lower the reserve below what the users have already borrowed.

pub fn update_users_limits( origin: OriginFor<T>, pot: T::PotId, new_fee_quota: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance, new_reserve_quota: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance, users: Vec<T::AccountId> ) -> DispatchResult

Update limits for a number of users in a single call. Only the sponsor can do this. If the sponsor is lowering their support, it can work only if the corresponding fee or reserve balance of all those users have enough available margin.

pub fn update_sponsorship_type( origin: OriginFor<T>, pot: T::PotId, sponsorship_type: T::SponsorshipType ) -> DispatchResult

Update the pot’s sponsorship type. Only the sponsor can do this. Emits PotSponsorshipTypeUpdated event when successful.