Crate pallet_sponsorship

Re-exports

pub use pallet::*;

pub use weights::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_sponsorship

Structs

ChargeSponsor

Require the sponsor to pay for their transactors.

PotDetails

A pot details a sponsorship and its limits. The remained fee/reserve quota of a pot is not withdrawn from the sponsor. So a valid pot does not guarantee that the sponsor has enough funds to cover the fees/reserves of the sponsored transactions.

PreDispatchSponsorCallData

UserDetails

Constants

STORAGE_VERSION

Type Aliases

Pre