Module pallet_sponsorship::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

create_pot

Create a new sponsorship pot and set its limits. The pot id shouldn’t be in use.

register_users

Register users for a pot and set the same limit for the list of them. Only pot sponsor can do this.

remove_pot

Allows the sponsor to remove the pot they have created themselves. The pot must not have any users. Users must have been removed prior to this call.

remove_users

Remove users from a pot. Only pot sponsor can do this. None of the specified users must have any reserved balance in their proxy accounts. User must be registered to be removable. Users receive the free balance in their proxy account back into their own accounts when they are removed.

sponsor_for

Sponsor me for the given call from the specified pot. The caller must be registered for the pot. The call must be consistent with the pot’s sponsorship type.

update_pot_limits

Update the pot details. Only the sponsor can do this. If the sponsor is lowering their support, it can work only if the corresponding fee or reserve balance has enough available margin. In other words, the sponsor cannot lower the limit for the fee below what users have already taken from the pot. Similarly, the sponsor cannot lower the reserve below what the users have already borrowed.

update_sponsorship_type

Update the pot’s sponsorship type. Only the sponsor can do this. Emits PotSponsorshipTypeUpdated event when successful.

update_users_limits

Update limits for a number of users in a single call. Only the sponsor can do this. If the sponsor is lowering their support, it can work only if the corresponding fee or reserve balance of all those users have enough available margin.